generateParameterFrequenciesTotal

Generate a set of frequencies with each parameter as per Saltelli, Tarantola & Chan, 1999.

Usage

generateParameterFrequenciesTotal(parameters, parameterNumber)

Arguments

parameters

List of SAParameter objects.

parameterNumber

An integer pointing to a specific parameter in the list of parameters.

Value

A vector of the frequencies associated with each parameter.